requirements.txt
aiogram~=3.0
environs~=9.0
redis
betterlogging 
openpyxl
psycopg2-binary
xlsxwriter

# # For enabling api:
# backoff
# ujson
# yarl

# # For PostgreSQL sqlalchemy + alembic:
# sqlalchemy~=2.0
# alembic~=1.0
# asyncpg
